[ADD] xodoo: init
``xodoo`` is collection of various scripts intended to be used for odoo.
It is based on ``click-odoo`` to connect and manage odoo.

Scripts
~~~~~~~

xodoo-migrate
-------------

Migration tool intended to run various migration scripts for Odoo. It
should be used when migration scripts directly on modules do not make
sense (e.g. need to generate some customer specific data that is not
directly related with any module).